@@ -1,9 +1,9 @@
 name: mnemonic
 description: Package for generating four letter Mnemonics
-version: 0.0.1
+version: 0.1.0
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
